Windows told Raku code it was not Windows, and every handle it returned came back as half of itself
Four faults between a Windows user and a working `is native` program, found
chasing a GUI module that asked the Windows loader for libgtk-3.so.0.

`$*DISTRO.is-win` was a hard-coded False — on every host. Rakudo DEFINES it as
the distro name being one of mswin32/mingw/msys/cygwin, and this engine
answered "windows", a name of its own invention, so the rule could never have
fired. `path-sep` said ':' there and `$*VM.config<osname>` said 'win32' where
MoarVM says 'MSWin32'. The reach is wider than modules: tools/install.raku
picks a native extension's file name off is-win, and so do t/run.raku,
tools/perf-guard.raku and tools/lib/Gate.rakumod — all of them believed they
were on POSIX.

The blind FFI path — the one Windows always takes, since it ships no libffi —
held eight integer arguments, and the Win32 API routinely goes past that:
CreateWindowExW takes twelve, CreateFontW fourteen. It is sixteen now; a
caller may always pass more arguments than the callee reads.

It also passed every value as C `long`, which is 32 bits under LLP64. Argument
slots, `is rw` slots, the return register and the callback trampolines all
truncated a pointer to its low half — and the return register did so with
libffi present too, which is how a real HMODULE (0x00007ff888990000) reached a
user as -2003238912, and why GetProcAddress then found nothing in it. All of
them are machine words now. On Win64 the same prototype cannot place a float
at all, because that ABI assigns argument registers by position rather than by
bank, so a floating-point argument there now says so instead of calling with
the value in the wrong place.

`nativecast(Pointer, &some-sub)` answered a Pointer to address 0. Nothing said
no: the cast fell through the type ladder to its "nothing matched" arm, so a
program that wrote that into a C struct handed the library a null function
pointer and failed somewhere else entirely. Rakudo refuses this cast; this now
says where a callback DOES become a C function pointer.

Gates: t/regression/ffi-wide-args-and-words.raku crosses sixteen integer
arguments, a pointer above 4 GB and a callback, once through libffi and once
with RAKUPP_FFI=0, with libffi as the oracle — it fails on the previous binary,
which is how it was checked. t/regression/nativecast-of-a-routine.raku pins the
diagnostic. t/windows/platform.raku checks the identity, CreateFontW's fourteen
arguments, a LoadLibraryW handle that still works as one in GetProcAddress, and
a 64-bit lParam through a callback; release.yml runs it on both Windows legs.
Local suite 820 → 821.

None of the Windows-side code has been executed here — this machine has no
Windows and no PowerShell, so CI is its first run.

+++ src/Interpreter.h
@@ -1022,7 +1022,10 @@ public:
     static long long ncRawAddr(const Value& v); // extract a raw pointer from a native value (0 if none)
     Value cglobal(const std::string& lib, const std::string& sym, const std::string& type); // C global variable
     std::string ncGuessLibraryName(const std::string& lib); // the file `is native(lib)` would dlopen
-    long runCallback(int slot, long a0, long a1, long a2, long a3, long a4, long a5); // NativeCall callback dispatch
+    // Machine words, not `long`: on Windows `long` is 32 bits, which truncated
+    // every pointer argument a C library handed a Raku callback.
+    long long runCallback(int slot, long long a0, long long a1, long long a2,
+                          long long a3, long long a4, long long a5); // NativeCall callback dispatch
     void runFfiClosure(void* closure, void* ret, void** args); // NativeCall ffi_closure dispatch
     // False on a thread the interpreter never entered — i.e. one the C library
     // made for itself. Such a thread has no lexical scope to run Raku in.
